Improve clean-up logic

+++ pkg/controllers/dynakube/connectioninfo/oneagent/reconciler.go
@@ -46,8 +46,10 @@ var NoOneAgentCommunicationHostsError = errors.New("no communication hosts for O
 
 func (r *reconciler) Reconcile(ctx context.Context) error {
 	if !r.dynakube.NeedAppInjection() && !r.dynakube.NeedsOneAgent() {
-		meta.RemoveStatusCondition(r.dynakube.Conditions(), oaConnectionInfoConditionType)
-		r.dynakube.Status.OneAgent.ConnectionInfoStatus = dynatracev1beta1.OneAgentConnectionInfoStatus{}
+		cond := meta.FindStatusCondition(*r.dynakube.Conditions(), oaConnectionInfoConditionType)
+		if cond == nil {
+			return nil // no condition == nothing is there to clean up
+		}
 
 		query := k8ssecret.NewQuery(ctx, r.client, r.apiReader, log)
 		err := query.Delete(r.dynakube.OneagentTenantSecret(), r.dynakube.Namespace)
@@ -55,6 +57,9 @@ func (r *reconciler) Reconcile(ctx context.Context) error {
 		if err != nil {
 			log.Error(err, "failed to clean-up OneAgent tenant-secret")
 		}
+
+		meta.RemoveStatusCondition(r.dynakube.Conditions(), oaConnectionInfoConditionType)
+		r.dynakube.Status.OneAgent.ConnectionInfoStatus = dynatracev1beta1.OneAgentConnectionInfoStatus{}
 
 		return nil // clean-up shouldn't cause a failure
 	}
